tests

+++ clinicadl/utils/maps_manager/maps_manager.py
@@ -2599,16 +2599,23 @@ class MapsManager:
         performance_path = (
             performance_dir / f"{data_group}_{self.mode}_level_prediction.tsv"
         )
-        results_df.to_csv(
-            performance_path, index=False, sep="\t", mode="a", header=False
-        )
+        if not performance_path.is_dir():
+            results_df.to_csv(performance_path, index=False, sep="\t", header=True)
+        else:
+            results_df.to_csv(
+                performance_path, index=False, sep="\t", mode="a", header=False
+            )
 
         metrics_path = performance_dir / f"{data_group}_{self.mode}_level_metrics.tsv"
         if metrics is not None:
             pd_metrics = pd.DataFrame(metrics).T
-            pd_metrics.to_csv(
-                metrics_path, index=False, sep="\t", mode="a", header=False
-            )
+
+            if not metrics_path.is_file():
+                pd_metrics.to_csv(metrics_path, index=False, sep="\t", header=True)
+            else:
+                pd_metrics.to_csv(
+                    metrics_path, index=False, sep="\t", mode="a", header=False
+                )
 
     def _ensemble_to_tsv(
         self,
@@ -3095,6 +3102,7 @@ class MapsManager:
             prediction_dir / f"{data_group}_{mode}_level_prediction.tsv",
             sep="\t",
         )
+        print(df)
         df.set_index(["participant_id", "session_id"], inplace=True, drop=True)
         return df
